requirements.txt
aniso8601==9.0.1
APScheduler==3.8.1
attrs==21.2.0
Authlib==0.15.5
backports.zoneinfo==0.2.1
BareNecessities==0.2.8
bcrypt==3.2.0
blinker==1.4
boto3==1.19.12
botocore==1.22.12
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
cryptography==35.0.0
csvalidate==1.1.1
dataclasses==0.6
dicttoxml==1.7.4
dnspython==2.1.0
docutils==0.18
dominate==2.6.0
email-validator==1.1.3
filelock==3.3.2
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-CSV==1.2.0
Flask-json-schema==0.0.5
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.1
importlib-resources==5.4.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.2
jmespath==0.10.0
jsonschema==4.2.1
Mail==2.1.0
manage==0.1.15
MarkupSafe==2.0.1
marshmallow==3.14.0
marshmallow-sqlalchemy==0.26.1
numpy==1.21.4
oauthlib==3.1.1
packaging==21.2
pandas==1.3.4
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.20
pycryptodome==3.11.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
secret==0.8
six==1.16.0
SQLAlchemy==1.4.26
tabulate==0.8.9
toml==0.10.2
tweepy==4.3.0
twython==3.9.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
visitor==0.1.3
Werkzeug==2.0.2
WTForms==2.3.3
xmltodict==0.12.0
zipp==3.6.0
